feat: add perf init output
