+++ lib/index.js
@@ -2,22 +2,22 @@ const { forkTsc, forkRun } = require('./process');
 const {
   parseArgs,
   debounce,
-  getIp,
   deleteFolderRecursive,
   readJSONCFile,
   copyFilesRecursive,
-  output,
-  colors,
   debug,
   getRelativeDir,
   triggerMessage,
   suffixMapping,
+  output,
+  colors,
 } = require('./util');
 const path = require('path');
 const { replaceTscAliasPaths } = require('tsc-alias');
 const chokidar = require('chokidar');
 const fs = require('fs');
 const { startProxyServer } = require('./proxy');
+const consoleOutput = require('./output');
 
 function run() {
   const { cmdPath, tscArgs, parentArgs, childArgs } = parseArgs(process.argv);
@@ -231,67 +231,18 @@ function run() {
           async (serverReportOption, isFirstCallback, during, debugUrl) => {
             if (isFirstCallback) {
               // 第一次启动把端口等信息打印出来
-              console.log('');
-              output(
-                `${colors.green('Node.js server')} ${colors.dim(
-                  'started in'
-                )} ${during} ms ${
-                  hasPaths
-                    ? colors.dim('and enable compile with ') + 'tsc-alias'
-                    : ''
-                }\n`
+              consoleOutput.renderServerFirstReady(
+                serverReportOption,
+                during,
+                hasPaths,
+                debugUrl
               );
-              if (serverReportOption && serverReportOption.port) {
-                const protocol = serverReportOption.ssl ? 'https' : 'http';
-                output(
-                  `${colors.green('➜')}  Local:    ${colors.cyan(
-                    `${protocol}://127.0.0.1:${colors.bright(
-                      serverReportOption.port
-                    )}${colors.cyan('/')}`
-                  )} `
-                );
-                const netWorkIp = getIp();
-                if (netWorkIp) {
-                  output(
-                    `${colors.green('➜')}  ${colors.dim(
-                      `Network:  ${protocol}://${netWorkIp}:${serverReportOption.port}/ `
-                    )}`
-                  );
-                }
-                if (debugUrl) {
-                  output(
-                    `${colors.green('➜')}  ${colors.dim(
-                      `Debugger: devtools://devtools/bundled/inspector.html?experiments=true&v8only=true&ws=${debugUrl.replace(
-                        'ws://',
-                        ''
-                      )}`
-                    )}`
-                  );
-                }
-                console.log('');
-              }
               if (proxyServer) {
                 proxyServer.start();
               }
               triggerMessage('server-first-ready');
             } else {
-              output('');
-              output(
-                `${colors.green('Node.js server')} ${colors.dim(
-                  'restarted in'
-                )} ${during} ms`
-              );
-              if (debugUrl) {
-                output(
-                  `${colors.green('➜')}  ${colors.dim(
-                    `Debugger: devtools://devtools/bundled/inspector.html?experiments=true&v8only=true&ws=${debugUrl.replace(
-                      'ws://',
-                      ''
-                    )}`
-                  )}`
-                );
-              }
-              output('');
+              consoleOutput.renderServerReady(during, debugUrl);
               triggerMessage('server-ready');
             }
           }
