+++ lib/process.js
@@ -2,8 +2,6 @@ const { fork, spawn } = require('child_process');
 const {
   filterFileChangedText,
   debug,
-  output,
-  colors,
   convertPosixToGnu,
 } = require('./util');
 const { CHILD_PROCESS_EXCEPTION_EXIT_CODE } = require('./constants');
@@ -11,6 +9,7 @@ const path = require('path');
 const fs = require('fs');
 // is windows
 const isWin = process.platform === 'win32';
+const consoleOutput = require('./output');
 
 /**
  *
@@ -104,7 +103,9 @@ const forkTsc = (tscArgs = [], options = {}) => {
 const forkRun = (runCmdPath, runArgs = [], options = {}) => {
   // 判断路径是否是包名还是路径
   if (!runCmdPath.startsWith('.') && !runCmdPath.startsWith('/')) {
-    runCmdPath = require.resolve(runCmdPath);
+    runCmdPath = require.resolve(runCmdPath, {
+      paths: [options.cwd || process.cwd()],
+    });
   }
   let runChild;
   let onServerReadyCallback;
@@ -125,6 +126,8 @@ const forkRun = (runCmdPath, runArgs = [], options = {}) => {
   const requiredExecArgv = gnuStyleParentArgs.filter(arg =>
     filterExecArgv.some(filterArg => arg.startsWith(filterArg))
   );
+
+  const isPerfInit = runArgs.includes('--perf-init');
 
   function innerFork(isFirstFork = false) {
     const startTime = Date.now();
@@ -157,10 +160,14 @@ const forkRun = (runCmdPath, runArgs = [], options = {}) => {
               Date.now() - startTime,
               currentDebugUrl
             ));
-          runChild.removeListener('message', onServerReady);
           lastBootstrapStatus = true;
         } else if (data.title === 'debug-url') {
           currentDebugUrl = data.debugUrl;
+        } else if (data.title === 'perf-init') {
+          if (isPerfInit && isFirstFork) {
+            // perf init
+            consoleOutput.renderPerfInit(data.data);
+          }
         }
       } catch (err) {
         console.error(err);
@@ -171,19 +178,7 @@ const forkRun = (runCmdPath, runArgs = [], options = {}) => {
     if (runArgs.includes('--keepalive')) {
       runChild.once('exit', code => {
         if (code === CHILD_PROCESS_EXCEPTION_EXIT_CODE) {
-          output(colors.red('*'.repeat(120)));
-          output(
-            `A ${colors.red(
-              `${colors.bright(
-                'Critical unhandledRejection or uncaughtException'
-              )}`
-            )} was detected and the process has exited automatically.`
-          );
-          output('Please make sure to handle it.');
-          output(
-            'The --keepalive parameter was enabled, we will do our best to ensure the process remains normal.'
-          );
-          output(colors.red('*'.repeat(120)));
+          consoleOutput.renderKeepAlive();
           if (lastBootstrapStatus === undefined || lastBootstrapStatus) {
             // 只有上一次启动成功了，才继续保活拉起，如果启动就失败了，就停止重启
             innerFork(false);
@@ -198,6 +193,7 @@ const forkRun = (runCmdPath, runArgs = [], options = {}) => {
   innerFork(true);
 
   const killRunningChild = async signal => {
+    runChild.removeAllListeners('message');
     if (isWin) {
       await new Promise(resolve => {
         if (!runChild || runChild.exitCode !== null) {
